requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django==2.1.3
python-dateutil==2.7.5
pytz==2018.7              # via django
six==1.11.0               # via python-dateutil
